requirements.txt
hydra-core
dash
requests
pandas
dash-bootstrap-components
pycoingecko
gunicorn
streamlit
Flask
flask-login
requests
flask-session
dash-auth
locust
cachelib
psycopg2-binary
nest-asyncio==1.5.8